backend/requirements.txt
streamlit
openpyxl
PyMuPDF
camelot-py[cv]
pandas
fastapi
uvicorn
python-multipart
